patched dangerous delete permission and enhanced admin list
git-svn-id: http://svn.ccnmtl.columbia.edu/ecomap/trunk@1330 1f418930-7ff8-0310-b8ad-c653122473bc

+++ ecomap/controller/__init__.py
@@ -374,8 +374,9 @@ class EcomapController(EcoControllerBase):
 
 
     def delete(self,confirm=""):
-        self.ecomap.destroySelf()
-        cherrypy.session['message'] = "deleted"
+        if cherrypy.session.get(UNI_PARAM,None) in ADMIN_USERS:
+            self.ecomap.destroySelf()
+            cherrypy.session['message'] = "deleted"
         return httptools.redirect("/myList")
 
     def flash(self):
